// src/app/nicolas/models/HojaVida.ts
export class HojaVida {
    cedula!: number;
    nombre?: String;
    apellido?: String;
    direccion?: String;
    correo?: String;
    fechaNacimiento?: String;
    twitter?: String;
    pinterest?: String;
    linkedIn?: String;
    hobbies?: String;
    trayectoria?: String;
    algoMas?: String;
    titulo?: String;
}